requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.1.7
django-admin-interface==0.24.2
django-ckeditor==6.7.0
django-colorfield==0.8.0
django-crispy-forms==1.13.0
django-jalali-date==1.0.1
django-js-asset==2.1.0
django-widget-tweaks==1.4.9
et-xmlfile==1.1.0
idna==3.4
jdatetime==4.1.0
kombu==5.3.2
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.0
Pillow==9.4.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
six==1.16.0
smsir-python==1.0.7
sorl-thumbnail==12.10.0
sqlparse==0.4.3
suds==1.1.2
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.6
xlwt==1.3.0
